Move part of WaitHandle to shared CoreLib
